added: bg click – close popup
Added the popupCloseHelper function, which handles the "Escape" key event and clicks on elements with the [popup-bg-close-trigger] attribute. When the "Escape" event occurs, the function finds all elements with the [popup-bg-close-button] attribute and clicks on each of them to close the corresponding pop-up windows. Additionally, the function also listens for click events on elements with the [popup-bg-close-trigger] attribute, and when such an event occurs, it finds the nearest element with the [popup-bg-close-button] attribute and clicks on it to close the respective pop-up window. This change improves the management of pop-up windows on the web page, making it more user-friendly for closing them.